// submodules/common-dto/api-client/main/models/ApiResponse_MainArtworkPassportPostResponseDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ApiResponse_MainArtworkPassportPostResponseDto = {
    /**
     * Indicates whether the request was successful
     */
    success: boolean;
    data: {
        /**
         * Url of generated report
         */
        url: string;
    };
};